Require parent app to call `use Anagrams`
Much of the behavior of Anagrams depends on the application environment
- eg, what dictionary we're using. But changing the application
  environment in a parent application doesn't trigger a recompile of
this library.

As a workaround, make the parent application define a module that uses
their application environment, and have it 'use Anagrams`, which defines
the various functions on their module directly. If they change their
environment, their module is recompiled, which redefines everything
appropriately.