Stop repeating work of subracting word from phrase
Previously, we were narrowing the dictionary at each recursion by
calling 'contains?', which called 'without'... and then threw away the
results. We later called 'without' again with the same input. That was
silly.

Now we just call 'without' and filter out what we can't use. Less work
should be faster, but is it?