Big changes, squashed. See summary.
"How lazy to squash so many unrelated changes into a commit", you say.

Yes. But the branch was so chaotic that this may be better. But yes.

Things:
- Add the concept of "process one job" - to explore one step further
  down the search tree, instead of recursing all the way down. This
  means we can...
- Run a job, get the resulting new jobs, put them in a queue, and
  repeat. We can use multiple processes to help process the queue.
- It also means we can (theoretically) stop after building N anagrams
  (though that isn't enabled yet)
- Unrelated except that it's about speed, start compiling all
  dictionaries up front
- Also unrelated, fix the way we handle custom lists of legal codepoints
- Changing the algo often causes ordering failures in tests, so fix
  those

The performance upshot is that we're now able to generate more than 2
million anagrams in about 7 seconds. :D

He doesn't care about credit, but for the record, I paired with the
brilliant Michael Gundlach on much of this, and couldn't have done it
without him.