Try to prioritize making anagrams with words early in the user wordlist
When iterating over the applicable wordlist for any remaining phrase,
and when generating jobs corresponding to those words, be sure that the
first ones correspond to the words earliest in the user's list.

This has the effect of prioritizing anagrams with words early in the
list, which means the user has more control over the kind of anagrams we
generate.

Eg, they can put political words, sports words, or scatalogical words at
the start of their list to increase the chances of finding anagrams
with those. This is especially relevant if they limit the number
produced - we want to produce the 100 (or whatever) "best" ones by their
definition of "best".

This isn't perfect because we do have some parallelism in the Queue, and
if more than one worker is working the queue, by definition only one of
them can be doing the highest priority job, and the other one may finish
first. I don't know a good solution for that, currently, but the large
batch size we're using means it shouldn't come into play until after
100k steps, so we should already be "zoomed in" quite a bit on anagrams
with words early in their list.